requirements.txt
streamlit
pandas
groq
google-generativeai
rapidfuzz
python-dotenv
openpyxl
